final rename of "reusable relay" classes
The changes are:

1. Rename Relay to Runner,
2. Rename RelayFactory to RunnerFactory,
3. Rename ReusableRelay to Relay,
4. Remove the existing RelayBuilder, and
5. Rename ReusableRelayBuilder to RelayBuilder.

The renames have the effect of making a Relay reusable by default, and the Runner the single-use dispatcher that the Relay uses.

Also added docblocks.
