--- src/RelayFactory.php
@@ -1,19 +1,0 @@
-<?php
-namespace Relay;
-
-class RelayFactory
-{
-    protected $queue;
-    protected $resolver;
-
-    public function __construct(array $queue, $resolver = null)
-    {
-        $this->queue = $queue;
-        $this->resolver = $resolver;
-    }
-
-    public function newInstance()
-    {
-        return new Relay($this->queue, $this->resolver);
-    }
-}
